requirements.txt
# requirements file 'pip install -r requirements.txt'

### BEGIN ###
appdirs==1.4.3
cycler==0.10.0
functools32==3.2.3.post2
matplotlib==2.0.2
MySQL-python==1.2.5
numpy==1.12.1
packaging==16.8
PyAudio==0.2.11
pydub==0.19.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
queuelib==1.4.2
queues==0.6.3
ruamel.ordereddict==0.4.9
ruamel.yaml==0.14.11
scipy==0.19.0
six==1.10.0
subprocess32==3.2.7
requests==2.14.2
Pillow==4.1.1

### END ###

